fix: Move activeID state to Accordion
When an FAQ group is opened, the previously
active FAQ group should become inactive. This
was not working because the state was not updated
on click on all the groups. When the activeID
state is moved up to the Accordion component,
changes to the state are seen by the child
AccordionGroup components, and are updated
accordingly.